validation/requirements.txt
fastapi
uvicorn
pandas
openpyxl
pyarrow
google-cloud-storage
google-cloud-logging